[frontend][core]: implemented useState hooks to manage the user responses

+++ backend/database/migrations/2024_01_16_204435_create_questionnaires_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('type_question_response');
             $table->string('time_question_response');
-            $table->integer('budget_question_response');
+            $table->float('budget_question_response');
             $table->unsignedBigInteger('room_id');
             $table->foreign('room_id')->references('id')->on('rooms');
             $table->timestamps();
